refactor: move file body-parser in loaders
